clean:rename the file to match its import name
